requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
altair
anyio
apache-airflow
apache-airflow-providers-common-compat
apache-airflow-providers-common-io
apache-airflow-providers-common-sql
apache-airflow-providers-fab
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-postgres
apache-airflow-providers-smtp
apache-airflow-providers-sqlite
apispec
argcomplete
asgiref
attrs
babel
beautifulsoup4
blinker
cachelib
cachetools
certifi
cffi
charset-normalizer
click
clickclick
colorama
colorlog
ConfigUpdater
connexion
cron-descriptor
croniter
cryptography
dash
dash-core-components
dash-html-components
dash-table
Deprecated
dill
dnspython
email_validator
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
frozendict
frozenlist
fsspec
gitdb
GitPython
google-re2
googleapis-common-protos
greenlet
grpcio
gunicorn
h11
html5lib
httpcore
httpx
idna
importlib_metadata
inflection
itsdangerous
Jinja2
jmespath
jsonschema
jsonschema-specifications
kafka-python
kafka-python-ng
lazy-object-proxy
limits
linkify-it-py
lockfile
lxml
Mako
markdown-it-py
MarkupSafe
marshmallow
marshmallow-oneofschema
marshmallow-sqlalchemy
mdit-py-plugins
mdurl
methodtools
more-itertools
multidict
multitasking
narwhals
nest-asyncio
numpy
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
packaging
pandas
pathspec
peewee
pendulum
pillow
platformdirs
plotly
pluggy
polars
prison
propcache
protobuf
psutil
psycopg2-binary
py4j
pyarrow
pycparser
pydeck
Pygments
PyJWT
pyspark
python-daemon
python-dateutil
python-nvd3
python-slugify
pytz
PyYAML
referencing
requests
requests-toolbelt
retrying
rfc3339-validator
rich
rich-argparse
rpds-py
selectolax
setproctitle
setuptools
six
smmap
sniffio
soupsieve
SQLAlchemy
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlparse
streamlit
selenium
tabulate
tenacity
termcolor
text-unidecode
time-machine
toml
tornado
typing_extensions
tzdata
uc-micro-py
universal_pathlib
urllib3
watchdog
webencodings
Werkzeug
wirerope
wrapt
WTForms
yarl
yfinance
zipp
python-dotenv
